requirements.txt
torch==2.1
streamlit
python-dotenv
PyPDF2
langchain